cfg file cleaner
