+++ .gitignore
@@ -3,3 +3,4 @@ data/flickr30k
 data/referit
 data/visual_genome
 tmp/
+weights/
