+++ code/utils.py
@@ -25,53 +25,6 @@ from yacs.config import CfgNode as CN
 # from maskrcnn_benchmark.utils.checkpoint import load_state_dict
 
 
-def update_from_dict(cfg: CN, dct: Dict[str, Any],
-                     key_maps: Dict[str, str] = None) -> CN:
-    """
-    Given original CfgNode (cfg) and input dictionary allows changing
-    the cfg with the updated dictionary values
-    Optional key_maps argument which defines a mapping between
-    same keys of the cfg node. Only used for convenience
-    Adapted from:
-    https://github.com/rbgirshick/yacs/blob/master/yacs/config.py#L219
-    """
-    # Original cfg
-    root = cfg
-
-    # Change the input dictionary using keymaps
-    # Now it is aligned with the cfg
-    full_key_list = list(dct.keys())
-    for full_key in full_key_list:
-        if full_key in key_maps:
-            cfg[full_key] = dct[full_key]
-            new_key = key_maps[full_key]
-            dct[new_key] = dct.pop(full_key)
-
-    # Convert the cfg using dictionary input
-    for full_key, v in dct.items():
-        if root.key_is_deprecated(full_key):
-            continue
-        if root.key_is_renamed(full_key):
-            root.raise_key_rename_error(full_key)
-        key_list = full_key.split(".")
-        d = cfg
-        for subkey in key_list[:-1]:
-            # Most important statement
-            assert subkey in d, f'key {full_key} doesnot exist'
-            d = d[subkey]
-
-        subkey = key_list[-1]
-        # Most important statement
-        assert subkey in d, f'key {full_key} doesnot exist'
-
-        value = cfg._decode_cfg_value(v)
-
-        assert isinstance(value, type(d[subkey]))
-        d[subkey] = value
-
-    return cfg
-
-
 def get_world_size():
     if not dist.is_available():
         return 1
